Move container definition to the pkg directory. Create default container instance.
